# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Brandyaction OS API
  version: 2.3.0
  description: 지식·업무·목표·콘텐츠·성과를 연결하는 브랜디액션 운영 API
servers:
  - url: https://brandyaction-os.vercel.app
security:
  - bearerAuth: []
paths:
...
  /api/v1/documents:
    get:
      summary: RLS가 허용하는 문서 목록
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { in: query, name: offset, schema: { type: integer, minimum: 0 } }
        - { in: query, name: statuses, schema: { type: string }, description: 쉼표로 구분한 상태 }
        - { in: query, name: folder, schema: { type: string } }
        - { in: query, name: owner, schema: { type: string, format: uuid } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        "200": { description: 문서 목록 }
    post:
      summary: 개인 초안 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentCreate"
      responses:
        "201": { description: 생성된 문서와 색인 상태 }
        "400": { $ref: "#/components/responses/ApiError" }
        "409": { $ref: "#/components/responses/ApiError", description: 같은 source/sourceRef 원본이 이미 존재함 }
    patch:
      summary: 기대 버전을 확인한 뒤 문서 수정. 정본 제목·본문은 변경 제안으로 저장
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DocumentCreate"
                - type: object
                  required: [id, expectedVersion]
                  properties:
                    id: { type: string, format: uuid }
                    expectedVersion: { type: integer, minimum: 1 }
      responses:
        "200": { description: 수정된 문서와 색인 상태 }
        "202": { description: 정본은 그대로이며 proposal 필드에 열린 변경 제안을 반환 }
        "409": { $ref: "#/components/responses/ApiError" }
    delete:
      summary: 문서를 삭제하지 않고 보관 상태로 전환
      parameters:
        - { in: query, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: 보관 완료 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase 사용자 JWT 또는 범위가 지정된 bos_pat 토큰(knowledge.read/knowledge.write/records.read/records.write)
  schemas:
...
    DocumentCreate:
      type: object
      required: [title, content]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        content: { type: string, minLength: 1, maxLength: 1500000 }
        folder: { type: string, maxLength: 160 }
        brand: { type: string, maxLength: 120 }
        team: { type: string, maxLength: 120 }
        tags: { type: array, maxItems: 30, items: { type: string, maxLength: 60 } }
        source: { type: string, maxLength: 40 }
        sourceRef: { type: [string, "null"], maxLength: 500 }
...
  responses:
    ApiError:
      description: 표준 오류 본문
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
                  details: {}